Refactor gender/blood type fields and update enums/models
Refactored Gender and BloodTypeId fields in DTOs and domain models from char/int to byte for improved type safety. Updated Gender and BloodType enums to use byte as the underlying type. Replaced Hospital/Department with HospitalId/DepartmentId in MedicalWorker models and DTOs. Adjusted repository seed data and controller logic to match new types and naming. Improved enum definitions and ensured consistency across models and DTOs.

+++ VKmfSoft_EHealth_API/Models/Domain/Patient/Patient.cs
@@ -11,8 +11,8 @@ namespace VKmfSoft_EHealth_API.Models.Domain.Patient
         public string? EmergencyContactName { get; set; }
         public string? EmergencyContactPhoneNumber { get; set; }
         public string? EmergencyContactDescription { get; set; }
-        public int BloodTypeId { get; set; }
+        public byte BloodTypeId { get; set; }
         public PatientHealthHistory? PatientHealthHistory { get; set; }
-
+       
     }
 }
